Use Accept / Content-Type headers properly
This change removes the necessity for most of the query parameter parsing,
allowing the *entire* query string for POST requests to be used for setting the
document title, which simplifies things. The server now expects this query
string to be submitted in URL percent-encoded format, and decodes it, which
means that the user can include spaces and special characters provided they are
willing to use a client-side library to encode as such.