Implement event subscription and notification (no in-page JS support yet)
Fully integrates event subscription and SSE handling into the page life cycle
and heartbeat management, as well as finishing out the new server paths to
handle event subscription and notification.

Yet to do: integrate with the page to actually create event listeners on page
elements, and send requests to the server when those listeners are triggered.