Use typed relative/absolute paths for specifying event subscriptions
This goes backward against the complicated selector/selected-based API that was
previously being explored. In cases with deep nesting of many fields, it may
mean marginally more transmitted data and work for the javascript engine, but
this is likely negligible in real use cases, and it is significantly simpler
this way.