Add new route ?resubscribe=<UUID> for seamless subscription changeover
A unique Content-Location header field is returned for each ?subscribe and now
?resubscribe request, indicating the location (...?resubscribe=<UUID>) where
resubscription requests should be issued. Upon a new subscription being POST-ed
to such a URL, the old stream will be terminated, and new events will flow to
the stream at the new URL.

Motivation: objects are created dynamically, and some of them may be event
targets. A mechanism is needed to allow subscriptions to be updated to suit,
without the client missing events or receiving duplicate events. The ?subscribe
API alone is not enough for this, because if streams are subscribed
simultaneously, it's not possible to compute overlap for certain (events may be
indistinguishable), and if they are subscribed sequentially, there may be events
that occur in the momentary gap.

This new API asks the *server* to atomically swap two streams for each other,
changing the subscription for the new stream.

To change subscriptions, a client:

1) POSTs a new subscription to the ?resubscribe=<UUID> path given in the
previous stream's Content-Location header field.
2) Finishes reading and processing the soon-to-close old event stream.
3) Resumes processing the new stream.

Since the UUID changes each time, this implements a token-passing scheme where
an agent currently reading a subscription stream is the only one with the
authority to terminate that stream by re-subscribing to it (since they are the
only one with the Content-Location information returned in that same stream's
response header). Attempts to ?resubscribe to non-existent (or even
previously-existent) streams will fail with an error status, meaning the API
must be used linearly, and is intended to be used sequentially by one thread of
execution.