Fix race condition leading to event listener detachment
The intentional yield point introduced in `setBodyTo` led to a race determined
by the event loop scheduler. Consider:

```javascript
setTimeout(() => {
    innerHTML(document.body, ...);
});
updateSubscription();
```

If the body of the callback executes first, all is well. If it executes second,
the subscription is updated on the document, *and then* the document is
re-written, which in some cases means nodes can be added/replaced
which *should* have a listener attached to them, but will not because the
content of the body was set after the listeners were re-processed.

This bug was first detectable on Chrome but could occasionally be elicited in
Firefox, and seemed more likely to happen with a fresh cache.

The solution was to bring the call to `updateSubscription` inside the timeout
callback and after the call to `innerHTML`, to ensure it operates over the *new*
DOM and not the old one:

```javascript
setTimeout(() => {
    innerHTML(document.body, ...);
    updateSubscription();
});
```

This commit also adds a more aggressive check to make sure we update the
subscription for the first time once the DOM is ready.