Significantly reduce lock contention during event processing
Previously, when we received an event, we locked the page updates stream
unconditionally, then perhaps sent a new subscription to the page if the receipt
of the event triggered an aggregate subscription change. This meant that every
single event required us to lock the update stream for the page, which meant
high frequency events and updates would contend with each other unnecessarily,
because in the vast majority of cases a new subscription does not need to be
issued to the page (this is only necessary when a subscriber disconnects or
connects). By inverting the order of conditionals in this one function, we allow
updates and events to run independently unless absolutely necessary.