Make sure clients get a refresh of page content on re-connect
If the myxine server is killed and then restarted, it's confusing for pages to
maintain their previous content without auto-refreshing to blank. In the general
case, we'd like any re-connecting client to receive a fresh copy of the page,
just in case their version is outdated. So, this commit sends a title and body
update to *all* clients every time any new client connects. This was already the
behavior for subscriptions, since the `Page::event_stream` function
unconditionally triggers a refresh of subscriptions for all clients.