Make use of resubscription messages once again
In initial prototypes, we always subscribed to all events -- but we already have
the infrastructure in place to send the correct list of events to the client, we
just weren't paying attention to it in the JavaScript. Now we are once more, so
only events that matter are sent back to the server.