Report HTML tags as lowercase rather than uppercase
Even though the HTML spec says that tags are case-insensitive and that the
tagName attribute should always return them in uppercase, users won't expect
this behavior since modern HTML is almost exclusively written in lowercase. As
such, we report the tagName in lowercase, to obviate the necessity for
case-insensitive comparison on the client side.