Add JavaScript evaluation escape hatch
The ?evaluate=... and ?evaluate (with data in body) routes now evaluate the
contents of those things as an expression/block (respectively) in the top-level
context of the browser. This is meant as a way to allow simple queries of the
document that otherwise would not be sent, such as window dimensions. Results
may get confusing if multiple browser windows are used (the future work on a
windowed mode may alleviate this).