Implement frame-numbering for page events internally
This change does not affect the public-facing API, but enables us to track from
which "frame" (i.e. generation of the DOM) each event originates, which will
open the path to a more precise per-frame API for discerning language bindings.

All raw uses of uuid::Uuid have been banished and replaced with the page::id
construct, which is indexed by the kind of "epoch" it represents. Id<E>s are
only *partially* ordered (equal-epoch Ids are incomparable unless their unique
components match).

Every page update sends a new frame id with it, which is then reported back when
an event occurs. The frame id reported is not necessarily the same frame id as
when the event was *attached* --- it is the one corresponding to the frame in
which the event *fired*.

In this changeset is also a mechanic to handle "transient" subscribers: that is,
sinks which subscribe only to events from one particular frame. This is not yet
surfaced to the public API, and has not been fully tested.