Add frame-bound event streams to the user-facing API
This change implements quite a few internal changes, to support a relatively
small but very useful API addition. Internal structural changes:

- Code cleanup: all explicit matches on the variant of `Content` are now
relegated to that module rather than occasionally used within `Page` in an
abstraction-breaking way. This encapsulation is not (yet) enforced, but could be
in future.
- Consolidate Page::set_title and Page::set_body into one Page::set_content, which
now takes an Option<Subscription> describing the event stream requested for this
frame update. This function required some intricate thinking about ordering an
locking, but is well-commented for the future.
- Subscribers now have an add_persistent_subscriber and add_transient_subscriber
methods, which both just wrap inserting a new Sink into the contained map, but
with keys indicating the type of subscriber
- Fix "thinko" about which comparison order to use when deciding whether to drop
a frame subscriber (frame sink should be dropped iff sink frame < event frame,
not the other way around)
- Rename internal-only API query parameters to unambiguously never clash in name
with external API parameters. There wasn't ambiguity... yet... but it's good
future-proofing.