Add ?refresh query parameter that forces reload
When POST-ing a new <body> with ?refresh present in the query string,
the page is instructed to do a full reload from the server after
performing the update. This is substantially slower than the incremental
diffing strategy used otherwise, but can help with "tricky" documents
like those containing changing embedded SVGs, which do not play nicely
with the diffHTML library currently used.