Remove frame tracking from myxine server
While frame-tracking initially seemed like a useful technique to prevent
out of date events from confusing UI logic, it ultimately proved to be a
dead end, because it imposed too great an implementation complexity
burden on the author of bindings, with dubious benefit, as
application-specific idempotency/validity guarantees can easily be
encoded in the model itself (that is, if a button is meant only to be
clicked when visible, the model can account for this by having it fail
to exist when it is not visible, and doing nothing for such spurious
events).

In the course of implementing frame-bound events, however, a number of
improvements to the library were implemented, and this branch contains
those changes, while removing the complexity of the frame-bound-events
feature that initially was the purpose of the branch.

The Python client, which never used the frame-bound events feature, does
not need to be updated. The Haskell client still needs to be reverted to
the old behavior before merging into master.