Use tokio Broadcast channel instead of hyper_usse server
This allows us to drop updates when the browser falls behind, which can
lead to snappier performance. Additionally, since hyper_usse is no
longer maintained, this is the first step in removing our use of
it. (The second step is to replace our use of hyper_usse::EventBuilder
with equivalent logic that does not depend on the crate.)