Eliminate aggregate subscriptions
In preparation for implementing long-polling, we need to remove the
notion of changing subscriptions on the page-side. This is because
long-polling requests, by their nature, are transient, and this means
there is no persistent concept of "total subscription on a page". In
this commit, we also remove the API for altering the contents of an
existing subscription stream, because in cases where this is desired,
the more full-featured long-polling API should instead be used.