Adapt direct-style Haskell client to long-polling API
This simplifies the use of the direct-style client. Previously, in order
to ensure the streaming connection stayed open, we would need to pass a
continuation to execute for each event. Now, because each event
corresponds to a separate request, we can have what was initially
desired:

```
events :: PageLocation -> EventList -> IO (IO PageEvent)
```