Resolve intermittent failure to set <title>
This collapses the set_title and set_content methods of Content to a
single method, set, which sets them both. This also simplifies the wire
protocol slightly to send a (title + body + refresh-mode) update in one
single event, rather than as a sequence of (title), (body +
refresh-mode) where the refresh mode was signaled out of band by the
event type.