Cancel pending JS query to the page when HTTP request is dropped
Previously, each JS query to the page would be indefinitely re-sent
until an answer was generated, even if the original consumer of the
query had disconnected. Now, when the HTTP connection corresponding to
an ?evaluate call is dropped, the corresponding query is removed from
the queries table for that page, and will no longer be (re)sent to
pages.