Add JavaScript evaluation to the Reactive monad
Now, there is the capability to evaluate JavaScript in event handlers.
This adds the MonadEval sealed typeclass to implement this addition.

This change also fixes a bug where all top-level event handlers were
firing multiple times if they were in a scope that contained other
nested events. Because of the way ConjMap-based dispatch works, an event
handler with an empty list of TargetFacts attached to it will match
every event, which means it will fire for things not just at the root
but everywhere. Nested documents mean that an N-nested document will
trigger the same top-level handler to fire N times. This is fixed by
having top-level handlers refer to a special "root" tag, rather than the
empty set of TargetFacts.