Add explicit window TargetFact
Instead of making a special root in the DOM, we now just route all
un-stopped bubbling events to the special Window TargetFact, and
anything registered on that is invoked last in the chain. This allows
the Reactive module to default to the Window target for top-level
listeners, making sure that all events at the top level -- even those
with no path at all -- are handled by any top-level registered
listeners.