Prevent default action for all listening events in the browser
Now, event.preventDefault() is called for all events we're listening to.
This means that checkboxes won't check themselves, etc., and removes a
source of statefulness from the view. This also means that the user will
have to implement that state themselves: but this is important that they
do, because otherwise there is hidden state, and that's bad!